--- app/src/App.css
@@ -1,29 +1,0 @@
-.App {
-  max-width: 700px;
-  margin: 50px auto 500px auto;
-}
-
-h2 {
-  margin: 25px 0 10px !important;
-}
-
-.floatleft {
-  float: left;
-}
-
-.floatright {
-  float: right;
-}
-
-.big-top-margin {
-  float: right;
-  margin: 300px 0 0 0;
-}
-
-.monaco-container {
-  position: absolute;
-  top: 50px;
-  left: 0;
-  right: 0;
-  bottom: 0;
-}
