+++ app/src/components/monaco-drawer.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+
+class AppMonacoDrawer extends React.Component {
+    render() {
+        return <div></div>
+    }
+}
