+++ app/src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 
 import * as serviceWorker from './serviceWorker';
 
-import App from './App';
+import App from './root';
 import { initPython } from './python/init';
 import './index.scss';
